server: add server persistence
Each time a command is queued, the instruction to save the current status of the games (completely identified by the map of game controllers) is put in the queue of the executor service of the game that triggered that command. To avoid concurrent accesses by two different game controllers to the backup file, the `save` method is `synchronized`.
This approach was preferred to running the `save` function with a periodic timer to ensure consistency, since the status of the server is saved only when a command has finished executing.

Co-authored-by: Mattia Vicenzotto <mattia.vicenzotto@mail.polimi.it>

Closes #63.