GameControllerFX: wait for scene creation before updating the board
The game did not correctly load the board and the bookshelf(ves) because the `update` method was called soon after `setScene("Game")`. The scene update, however, was delayed by the JavaFX thread since it was declared inside a `Platform.runLater()` block. Added a call to a util method which should wait for the previous `runLater()` to be completed.

Closes #87